rename component: Event -> EventSelector
