+++ frontend/src/components/EventSelector.vue
